Created a pattern module with a Pattern class, for encapsulating a parse tree (precompiled pattern).  This allows for reuse of the pattern without needing to reparse it each time.
